+++ app/menuPrincipal/financeiro/index.tsx
@@ -1,4 +1,4 @@
-import MenuLateral from '../../components/menuLateral'
+import MenuLateral from '../../components/menuInferior'
 
 export default function Financeiro() {
   return (
